+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - New HTML parse filter: "json-parse" (useful for parsing json-ld for instance)
 - New HTML parse extractor: "count", counts the number of elements matching the selector (needs an array schema definition)
 - New "csv-file" loader
+- New bulk processor (TODO: update README)
 - New demos
 
 ### Changed